pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qqmail-cli"
version = "0.2.0"
description = "腾讯企业邮箱 IMAP 命令行客户端"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.3.2",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11",
]

[project.scripts]
qqmail = "qqmail_cli.main:cli"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
